+++ app/src/main/java/org/simple/clinic/cvdrisk/NonLabBasedCVDRiskInput.kt
@@ -1,0 +1,13 @@
+package org.simple.clinic.cvdrisk
+
+import org.simple.clinic.medicalhistory.Answer
+import org.simple.clinic.patient.Gender
+
+data class NonLabBasedCVDRiskInput(
+    val gender: Gender,
+    val age: Int,
+    val systolic: Int,
+    val isSmoker: Answer = Answer.Unanswered,
+    val bmi: Float? = null,
+)
+