+++ app/src/test/java/org/simple/clinic/summary/PatientSummaryEffectHandlerTest.kt
@@ -13,10 +13,11 @@ import org.mockito.kotlin.whenever
 import org.simple.clinic.TestData
 import org.simple.clinic.bloodsugar.BloodSugarRepository
 import org.simple.clinic.bp.BloodPressureRepository
-import org.simple.clinic.cvdrisk.CVDRiskCalculator
+import org.simple.clinic.cvdrisk.calculator.NonLabBasedCVDRiskCalculator
 import org.simple.clinic.cvdrisk.CVDRiskRange
 import org.simple.clinic.cvdrisk.CVDRiskRepository
 import org.simple.clinic.cvdrisk.StatinInfo
+import org.simple.clinic.cvdrisk.calculator.LabBasedCVDRiskCalculator
 import org.simple.clinic.drugs.DiagnosisWarningPrescriptions
 import org.simple.clinic.drugs.PrescriptionRepository
 import org.simple.clinic.facility.FacilityRepository
@@ -79,7 +80,8 @@ class PatientSummaryEffectHandlerTest {
       htnPrescriptions = listOf("amlodipine"),
       diabetesPrescriptions = listOf("metformin")
   )
-  private val cvdRiskCalculator = CVDRiskCalculator { TestData.nonLabBasedCVDRiskCalculationSheet() }
+  private val nonLabBasedCVDRiskCalculator = NonLabBasedCVDRiskCalculator { TestData.nonLabBasedCVDRiskCalculationSheet() }
+  private val labBasedCVDRiskCalculator = LabBasedCVDRiskCalculator { TestData.labBasedCVDRiskCalculationSheet() }
 
   private val effectHandler = PatientSummaryEffectHandler(
       clock = clock,
@@ -92,6 +94,8 @@ class PatientSummaryEffectHandlerTest {
       bloodSugarRepository = bloodSugarRepository,
       dataSync = dataSync,
       medicalHistoryRepository = medicalHistoryRepository,
+      cvdRiskRepository = cvdRiskRepository,
+      patientAttributeRepository = patientAttributeRepository,
       country = TestData.country(),
       currentUser = { user },
       currentFacility = { facility },
@@ -101,10 +105,9 @@ class PatientSummaryEffectHandlerTest {
       prescriptionRepository = prescriptionRepository,
       cdssPilotFacilities = { emptyList() },
       diagnosisWarningPrescriptions = { diagnosisWarningPrescriptions },
-      cvdRiskRepository = cvdRiskRepository,
+      nonLabBasedCVDRiskCalculator = nonLabBasedCVDRiskCalculator,
+      labBasedCVDRiskCalculator = labBasedCVDRiskCalculator,
       viewEffectsConsumer = viewEffectHandler::handle,
-      cvdRiskCalculator = cvdRiskCalculator,
-      patientAttributeRepository = patientAttributeRepository,
   )
   private val testCase = EffectHandlerTestCase(effectHandler.build())
 
@@ -138,6 +141,8 @@ class PatientSummaryEffectHandlerTest {
         bloodSugarRepository = bloodSugarRepository,
         dataSync = dataSync,
         medicalHistoryRepository = medicalHistoryRepository,
+        cvdRiskRepository = cvdRiskRepository,
+        patientAttributeRepository = patientAttributeRepository,
         country = bangladesh,
         currentUser = { user },
         currentFacility = { facility },
@@ -146,11 +151,10 @@ class PatientSummaryEffectHandlerTest {
         teleconsultationFacilityRepository = teleconsultFacilityRepository,
         prescriptionRepository = prescriptionRepository,
         cdssPilotFacilities = { emptyList() },
-        viewEffectsConsumer = viewEffectHandler::handle,
         diagnosisWarningPrescriptions = { diagnosisWarningPrescriptions },
-        cvdRiskRepository = cvdRiskRepository,
-        cvdRiskCalculator = cvdRiskCalculator,
-        patientAttributeRepository = patientAttributeRepository,
+        nonLabBasedCVDRiskCalculator = nonLabBasedCVDRiskCalculator,
+        labBasedCVDRiskCalculator = labBasedCVDRiskCalculator,
+        viewEffectsConsumer = viewEffectHandler::handle,
     )
     val testCase = EffectHandlerTestCase(effectHandler.build())
     val registeredFacilityUuid = UUID.fromString("1b359ec9-02e2-4f50-bebd-6001f96df57f")
@@ -201,6 +205,8 @@ class PatientSummaryEffectHandlerTest {
         bloodSugarRepository = bloodSugarRepository,
         dataSync = dataSync,
         medicalHistoryRepository = medicalHistoryRepository,
+        cvdRiskRepository = cvdRiskRepository,
+        patientAttributeRepository = patientAttributeRepository,
         country = bangladesh,
         currentUser = { user },
         currentFacility = { facility },
@@ -210,10 +216,9 @@ class PatientSummaryEffectHandlerTest {
         prescriptionRepository = prescriptionRepository,
         cdssPilotFacilities = { emptyList() },
         diagnosisWarningPrescriptions = { diagnosisWarningPrescriptions },
+        nonLabBasedCVDRiskCalculator = nonLabBasedCVDRiskCalculator,
+        labBasedCVDRiskCalculator = labBasedCVDRiskCalculator,
         viewEffectsConsumer = viewEffectHandler::handle,
-        cvdRiskRepository = cvdRiskRepository,
-        cvdRiskCalculator = cvdRiskCalculator,
-        patientAttributeRepository = patientAttributeRepository,
     )
     val testCase = EffectHandlerTestCase(effectHandler.build())
     val patient = TestData.patient(patientUuid)
@@ -633,6 +638,8 @@ class PatientSummaryEffectHandlerTest {
         bloodSugarRepository = bloodSugarRepository,
         dataSync = dataSync,
         medicalHistoryRepository = medicalHistoryRepository,
+        cvdRiskRepository = cvdRiskRepository,
+        patientAttributeRepository = patientAttributeRepository,
         country = TestData.country(),
         currentUser = { user },
         currentFacility = { facility },
@@ -642,10 +649,9 @@ class PatientSummaryEffectHandlerTest {
         prescriptionRepository = prescriptionRepository,
         cdssPilotFacilities = { cdssPilotFacilities },
         diagnosisWarningPrescriptions = { diagnosisWarningPrescriptions },
-        cvdRiskRepository = cvdRiskRepository,
+        nonLabBasedCVDRiskCalculator = nonLabBasedCVDRiskCalculator,
+        labBasedCVDRiskCalculator = labBasedCVDRiskCalculator,
         viewEffectsConsumer = viewEffectHandler::handle,
-        cvdRiskCalculator = cvdRiskCalculator,
-        patientAttributeRepository = patientAttributeRepository,
     )
     val testCase = EffectHandlerTestCase(effectHandler = effectHandler.build())
 
@@ -875,7 +881,7 @@ class PatientSummaryEffectHandlerTest {
   }
 
   @Test
-  fun `when calculate cvd risk effect is received, then calculate cvd risk`() {
+  fun `when calculate non lab based cvd risk effect is received, then calculate non lab based cvd risk`() {
     //given
     val patient = TestData.patient(
         uuid = patientUuid,
@@ -906,13 +912,50 @@ class PatientSummaryEffectHandlerTest {
         patientUuid = patientUuid,
     )) doReturn null
     whenever(cvdRiskRepository.getCVDRiskImmediate(patientUuid)) doReturn
-       cvdRisk
+        cvdRisk
 
     //when
     testCase.dispatch(CalculateNonLabBasedCVDRisk(patient = patient))
 
     //then
     testCase.assertOutgoingEvents(CVDRiskCalculated(cvdRisk, CVDRiskRange(6, 6)))
+  }
+
+  @Test
+  fun `when calculate lab based cvd risk effect is received, then calculate lab based cvd risk`() {
+    //given
+    val patient = TestData.patient(
+        uuid = patientUuid,
+        gender = Gender.Male,
+        status = PatientStatus.Active,
+        patientAgeDetails = PatientAgeDetails(
+            ageValue = 40,
+            ageUpdatedAt = Instant.parse("2018-01-01T00:00:00Z"),
+            dateOfBirth = null,
+        )
+    )
+    val medicalHistory = TestData.medicalHistory(isSmoking = Yes)
+    val bloodPressure = TestData.bloodPressureMeasurement(
+        UUID.fromString("3e8c246f-91b9-4f8c-81fe-91b67ac0a2d5"),
+        systolic = 130,
+        patientUuid = patientUuid
+    )
+    val bloodPressures = listOf(bloodPressure)
+    val cvdRisk = TestData.cvdRisk(riskScore = CVDRiskRange(27, 27))
+
+    whenever(bloodPressureRepository.newestMeasurementsForPatientImmediate(patientUuid = patientUuid, limit = 1)) doReturn bloodPressures
+    whenever(medicalHistoryRepository.historyForPatientOrDefaultImmediate(
+        patientUuid = patientUuid,
+        defaultHistoryUuid = uuidGenerator.v4()
+    )) doReturn medicalHistory
+    whenever(cvdRiskRepository.getCVDRiskImmediate(patientUuid)) doReturn
+        cvdRisk
+
+    //when
+    testCase.dispatch(CalculateLabBasedCVDRisk(patient = patient))
+
+    //then
+    testCase.assertOutgoingEvents(CVDRiskCalculated(cvdRisk, CVDRiskRange(7, 14)))
   }
 
   @Test